pyproject.toml
[tool.poetry]
name = "django-test"
version = "0.1.0"
description = ""
authors = ["gnthibault <gnthibault@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
django-debug-toolbar = "^4.4.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
